Refactor modules and update tests
