+++ src/massconfigmerger/vpn_merger.py
@@ -25,31 +25,36 @@ Expected Output: 800k-1.2M+ tested and sorted configs
 """
 
 import asyncio
-import base64
-import binascii
-import csv
-import hashlib
-import json
 import logging
 import random
 import re
-import yaml
 import ssl
 import sys
 import time
 import socket
-from dataclasses import dataclass
+import json
+import base64
+import binascii
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Set, Tuple, Union, cast, Callable, Awaitable
 
-from .constants import SOURCES_FILE
-from urllib.parse import urlparse, parse_qs, urlencode, urlunparse, parse_qsl
+from urllib.parse import urlparse, parse_qs
 from tqdm import tqdm
-from .clash_utils import config_to_clash_proxy, flag_emoji, build_clash_config
 
-from .config import Settings, load_config
-from .tester import NodeTester
+from .source_fetcher import (
+    fetch_text,
+    parse_first_configs,
+    UnifiedSources,
+    AsyncSourceFetcher,
+)
+from .result_processor import (
+    CONFIG,
+    ConfigResult,
+    EnhancedConfigProcessor,
+)
+from .output_writer import generate_comprehensive_outputs, EXCLUDE_REGEXES as OUTPUT_EXCLUDE_REGEXES
+from .clash_utils import config_to_clash_proxy, flag_emoji, build_clash_config
 
 try:
     import aiohttp
@@ -86,476 +91,9 @@ except ImportError as exc:
 # CONFIGURATION & SETTINGS
 # ============================================================================
 
-DEFAULT_CONFIG_FILE = Path(__file__).resolve().with_name("config.yaml")
-try:
-    CONFIG = load_config(DEFAULT_CONFIG_FILE)
-except ValueError:
-    CONFIG = Settings()
 
-# Compiled regular expressions from --exclude-pattern
-EXCLUDE_REGEXES: List[re.Pattern] = []
-
-# Regex patterns for extracting configuration links during pre-flight checks
-PROTOCOL_RE = re.compile(
-    r"(?:"
-    r"vmess|vless|reality|ssr?|trojan|hy2|hysteria2?|tuic|"
-    r"shadowtls|juicity|naive|brook|wireguard|"
-    r"socks5|socks4|socks|http|https|grpc|ws|wss|"
-    r"tcp|kcp|quic|h2"
-    r")://\S+",
-    re.IGNORECASE,
-)
-BASE64_RE = re.compile(r"^[A-Za-z0-9+/=_-]+$")
-MAX_DECODE_SIZE = 256 * 1024  # 256 kB
-
-
-async def fetch_text(
-    session: aiohttp.ClientSession,
-    url: str,
-    timeout: int = 10,
-    *,
-    retries: int = 3,
-    base_delay: float = 1.0,
-    jitter: float = 0.1,
-) -> str | None:
-    """Fetch text content with retries."""
-    parsed = urlparse(url)
-    if not parsed.scheme or not parsed.netloc:
-        logging.debug("fetch_text invalid url: %s", url)
-        return None
-
-    attempt = 0
-    while attempt < retries:
-        try:
-            async with session.get(url, timeout=aiohttp.ClientTimeout(total=timeout)) as resp:
-                if resp.status == 200:
-                    return await resp.text()
-                if 400 <= resp.status < 500 and resp.status != 429:
-                    logging.debug("fetch_text non-retry status %s on %s", resp.status, url)
-                    return None
-                if not (500 <= resp.status < 600 or resp.status == 429):
-                    logging.debug(
-                        "fetch_text non-transient status %s on %s", resp.status, url
-                    )
-                    return None
-        except (aiohttp.ClientError, asyncio.TimeoutError) as exc:
-            logging.debug("fetch_text error on %s: %s", url, exc)
-
-        attempt += 1
-        if attempt >= retries:
-            break
-        delay = base_delay * 2 ** (attempt - 1)
-        await asyncio.sleep(delay + random.uniform(0, jitter))
-    return None
-
-
-def parse_first_configs(text: str, limit: int = 5) -> List[str]:
-    """Extract up to ``limit`` configuration links from ``text``."""
-    configs: List[str] = []
-    for line in text.splitlines():
-        line = line.strip()
-        matches = PROTOCOL_RE.findall(line)
-        if matches:
-            for m in matches:
-                configs.append(m)
-                if len(configs) >= limit:
-                    return configs
-            continue
-        if BASE64_RE.match(line):
-            if len(line) > MAX_DECODE_SIZE:
-                continue
-            try:
-                padded = line + "=" * (-len(line) % 4)
-                decoded = base64.urlsafe_b64decode(padded).decode()
-                for m in PROTOCOL_RE.findall(decoded):
-                    configs.append(m)
-                    if len(configs) >= limit:
-                        return configs
-            except (binascii.Error, UnicodeDecodeError):
-                continue
-    return configs
-
-# ============================================================================
-# COMPREHENSIVE SOURCE COLLECTION (ALL UNIFIED SOURCES)
-# ============================================================================
-
-class UnifiedSources:
-    """Load VPN subscription sources from an external file."""
-
-    DEFAULT_FILE = SOURCES_FILE
-    FALLBACK_SOURCES = [
-        "https://raw.githubusercontent.com/barry-far/V2ray-config/main/Sub1.txt",
-        "https://raw.githubusercontent.com/ssrsub/ssr/master/v2ray",
-        "https://raw.githubusercontent.com/mahdibland/V2RayAggregator/master/sub/splitted/vmess.txt",
-    ]
-
-    @classmethod
-    def load(cls, path: Optional[Union[str, Path]] = None) -> List[str]:
-        file_path = Path(path) if path else cls.DEFAULT_FILE
-        if file_path.exists():
-            with file_path.open() as f:
-                return [line.strip() for line in f if line.strip()]
-        logging.warning("sources file not found: %s; using fallback list", file_path)
-        return cls.FALLBACK_SOURCES
-
-    @classmethod
-    def get_all_sources(cls, path: Optional[Union[str, Path]] = None) -> List[str]:
-        """Return list of sources from file or fallback."""
-        return cls.load(path)
-
-@dataclass
-class ConfigResult:
-    """Enhanced config result with testing metrics."""
-    config: str
-    protocol: str
-    host: Optional[str] = None
-    port: Optional[int] = None
-    ping_time: Optional[float] = None
-    is_reachable: bool = False
-    source_url: str = ""
-    country: Optional[str] = None
-
-class EnhancedConfigProcessor:
-    """Advanced configuration processor with comprehensive testing capabilities."""
-    
-    MAX_DECODE_SIZE = 256 * 1024  # 256 kB safety limit for base64 payloads
-
-    def __init__(self):
-        self.tester = NodeTester(CONFIG)
-
-    def _normalize_url(self, config: str) -> str:
-        """Return canonical URL with sorted query params and no fragment."""
-        parsed = urlparse(config)
-
-        # Canonicalize query parameters
-        query_pairs = parse_qsl(parsed.query, keep_blank_values=True)
-        sorted_query = urlencode(sorted(query_pairs), doseq=True)
-
-        scheme = parsed.scheme.lower()
-
-        # Canonicalize base64 payloads used by vmess/vless links so that key
-        # ordering in the JSON does not affect the final hash.
-        if scheme in {"vmess", "vless"}:
-            payload = parsed.netloc or parsed.path.lstrip("/")
-            if payload:
-                try:
-                    padded = payload + "=" * (-len(payload) % 4)
-                    decoded = base64.b64decode(padded).decode("utf-8", "ignore")
-                    data = json.loads(decoded)
-                    canonical_json = json.dumps(data, sort_keys=True)
-                    payload = base64.b64encode(canonical_json.encode()).decode().rstrip("=")
-                    parsed = parsed._replace(netloc=payload, path="")
-                except (binascii.Error, UnicodeDecodeError, json.JSONDecodeError):
-                    pass
-
-        return urlunparse(parsed._replace(query=sorted_query, fragment=""))
-        
-    def extract_host_port(self, config: str) -> Tuple[Optional[str], Optional[int]]:
-        """Extract host and port from configuration for testing."""
-        try:
-            if config.startswith(("vmess://", "vless://")):
-                try:
-                    json_part = config.split("://", 1)[1]
-                    decoded_bytes = base64.b64decode(json_part)
-                    if len(decoded_bytes) > self.MAX_DECODE_SIZE:
-                        return None, None
-                    decoded = decoded_bytes.decode("utf-8", "ignore")
-                    data = json.loads(decoded)
-                    host = data.get("add") or data.get("host")
-                    port = data.get("port")
-                    return host, int(port) if port else None
-                except (binascii.Error, UnicodeDecodeError, json.JSONDecodeError, ValueError) as exc:
-                    logging.debug("extract_host_port vmess failed: %s", exc)
-
-            if config.startswith("ssr://"):
-                try:
-                    after = config.split("://", 1)[1].split("#", 1)[0]
-                    padded = after + "=" * (-len(after) % 4)
-                    decoded = base64.urlsafe_b64decode(padded).decode()
-                    host_part = decoded.split("/", 1)[0]
-                    parts = host_part.split(":")
-                    if len(parts) < 2:
-                        return None, None
-                    host, port = parts[0], parts[1]
-                    return host or None, int(port)
-                except (binascii.Error, UnicodeDecodeError, ValueError) as exc:
-                    logging.debug("extract_host_port ssr failed: %s", exc)
-            
-            # Parse URI-style configs
-            parsed = urlparse(config)
-            if parsed.hostname and parsed.port:
-                return parsed.hostname, parsed.port
-                
-            # Extract from @ notation
-            match = re.search(r"@([^:/?#]+):(\d+)", config)
-            if match:
-                return match.group(1), int(match.group(2))
-                
-        except (ValueError, UnicodeError, binascii.Error) as exc:
-            logging.debug("extract_host_port failed: %s", exc)
-        return None, None
-    
-    def create_semantic_hash(self, config: str) -> str:
-        """Create semantic hash for intelligent deduplication."""
-        parsed = urlparse(config)
-        normalized_config = self._normalize_url(config)
-
-        host, port = self.extract_host_port(normalized_config)
-        identifier = None
-
-        scheme = parsed.scheme.lower()
-
-        if scheme in ("vmess", "vless"):
-            try:
-                after_scheme = normalized_config.split("://", 1)[1].split("?", 1)[0]
-                if parsed.username:
-                    identifier = parsed.username
-                else:
-                    padded = after_scheme + "=" * (-len(after_scheme) % 4)
-                    decoded = base64.b64decode(padded).decode("utf-8", "ignore")
-                    data = json.loads(decoded)
-                    # Re-serialize with sorted keys for consistent hashing
-                    json.dumps(data, sort_keys=True)
-                    identifier = data.get("id") or data.get("uuid") or data.get("user")
-            except (binascii.Error, UnicodeDecodeError, json.JSONDecodeError, ValueError) as exc:
-                logging.debug("semantic_hash vmess failed: %s", exc)
-        elif scheme == "trojan":
-            try:
-                parsed = urlparse(normalized_config)
-                if parsed.username or parsed.password:
-                    identifier = parsed.username or ""
-                    if parsed.password is not None:
-                        if identifier:
-                            identifier += f":{parsed.password}"
-                        else:
-                            identifier = parsed.password
-                else:
-                    identifier = None
-            except ValueError as exc:
-                logging.debug("semantic_hash trojan failed: %s", exc)
-        elif scheme in ("ss", "shadowsocks"):
-            try:
-                parsed = urlparse(normalized_config)
-                if parsed.username and parsed.password:
-                    identifier = parsed.password
-                else:
-                    base = normalized_config.split("://", 1)[1]
-                    base = base.split("?", 1)[0]
-                    padded = base + "=" * (-len(base) % 4)
-                    decoded = base64.b64decode(padded).decode("utf-8", "ignore")
-                    before_at = decoded.split("@", 1)[0]
-                    method, password = before_at.split(":", 1)
-                    identifier = password
-            except (binascii.Error, UnicodeDecodeError, ValueError) as exc:
-                logging.debug("semantic_hash ss failed: %s", exc)
-
-        if host and port:
-            key = f"{host}:{port}"
-            if identifier:
-                key = f"{identifier}@{key}"
-        else:
-            key = normalized_config.strip()
-        return hashlib.sha256(key.encode()).hexdigest()[:16]
-    
-    async def test_connection(self, host: str, port: int) -> Optional[float]:
-        """Test connection and measure response time using :class:`NodeTester`."""
-        return await self.tester.test_connection(host, port)
-
-    async def lookup_country(self, host: str) -> Optional[str]:
-        """Return ISO country code for ``host`` using :class:`NodeTester`."""
-        return await self.tester.lookup_country(host)
-
-    def categorize_protocol(self, config: str) -> str:
-        """Categorize configuration by protocol."""
-        protocol_map = {
-            "vmess://": "VMess",
-            "vless://": "VLESS", 
-            "ss://": "Shadowsocks",
-            "ssr://": "ShadowsocksR",
-            "trojan://": "Trojan",
-            "hy2://": "Hysteria2",
-            "hysteria2://": "Hysteria2",
-            "hysteria://": "Hysteria",
-            "tuic://": "TUIC",
-            "reality://": "Reality",
-            "naive://": "Naive",
-            "juicity://": "Juicity",
-            "wireguard://": "WireGuard",
-            "shadowtls://": "ShadowTLS",
-            "brook://": "Brook",
-        }
-        
-        for prefix, protocol in protocol_map.items():
-            if config.startswith(prefix):
-                return protocol
-
-        return "Other"
-
-    def apply_tuning(self, config: str) -> str:
-        """Apply mux and smux parameters to URI-style configs."""
-        try:
-            if "//" not in config or config.startswith("vmess://"):
-                return config
-            parsed = urlparse(config)
-            if not parsed.scheme:
-                return config
-            params = parse_qs(parsed.query)
-            if CONFIG.mux_concurrency > 0:
-                params["mux"] = [str(CONFIG.mux_concurrency)]
-            if CONFIG.smux_streams > 0:
-                params["smux"] = [str(CONFIG.smux_streams)]
-            new_query = urlencode(params, doseq=True)
-            return urlunparse(parsed._replace(query=new_query))
-        except ValueError as exc:
-            logging.debug("apply_tuning failed: %s", exc)
-            return config
-
-# ============================================================================
-# ASYNC SOURCE FETCHER WITH COMPREHENSIVE TESTING
-# ============================================================================
-
-class AsyncSourceFetcher:
-    """Async source fetcher with comprehensive testing and availability checking."""
-
-    def __init__(
-        self,
-        processor: EnhancedConfigProcessor,
-        seen_hashes: Set[str],
-        hash_lock: Optional[asyncio.Lock] = None,
-        history_callback: Optional[
-            Callable[[str, bool, Optional[float]], Awaitable[None]]
-        ] = None,
-    ):
-        self.processor = processor
-        self.session: Optional[aiohttp.ClientSession] = None
-        self.seen_hashes = seen_hashes
-        self.hash_lock = hash_lock or asyncio.Lock()
-        self.history_callback = history_callback
-        
-    async def test_source_availability(self, url: str) -> bool:
-        """Test if a source URL is available (returns 200 status)."""
-        session = self.session
-        if session is None or getattr(session, "loop", None) is not asyncio.get_running_loop():
-            session = aiohttp.ClientSession()
-            close_temp = True
-        else:
-            close_temp = False
-        try:
-            timeout = aiohttp.ClientTimeout(total=10)
-            async with session.head(url, timeout=timeout, allow_redirects=True) as response:
-                status = response.status
-                if status == 200:
-                    return True
-                if 400 <= status < 500:
-                    async with session.get(
-                        url,
-                        headers={**CONFIG.headers, 'Range': 'bytes=0-0'},
-                        timeout=timeout,
-                        allow_redirects=True,
-                    ) as get_resp:
-                        return get_resp.status in (200, 206)
-                return False
-        except (aiohttp.ClientError, asyncio.TimeoutError) as exc:
-            logging.debug("availability check failed for %s: %s", url, exc)
-            return False
-        finally:
-            if close_temp:
-                await session.close()
-        
-    async def fetch_source(self, url: str) -> Tuple[str, List[ConfigResult]]:
-        """Fetch single source with comprehensive testing and deduplication."""
-        session = self.session
-        use_temp = session is None or getattr(session, "loop", None) is not asyncio.get_running_loop()
-        if use_temp:
-            session = aiohttp.ClientSession()
-        for attempt in range(CONFIG.max_retries):
-            try:
-                timeout = aiohttp.ClientTimeout(total=CONFIG.request_timeout)
-                async with session.get(url, headers=CONFIG.headers, timeout=timeout) as response:
-                    if response.status != 200:
-                        continue
-
-                    content = await response.text()
-                    if not content.strip():
-                        return url, []
-                    
-                    # Enhanced Base64 detection and decoding
-                    try:
-                        # Check if content looks like base64
-                        if not any(char in content for char in '\n\r') and len(content) > 100:
-                            decoded = base64.b64decode(content).decode("utf-8", "ignore")
-                            if decoded.count("://") > content.count("://"):
-                                content = decoded
-                    except (binascii.Error, UnicodeDecodeError) as exc:  # noqa: E722
-                        logging.debug("Base64 decode failed: %s", exc)
-                    
-                    # Extract and process configs
-                    lines = [line.strip() for line in content.splitlines() if line.strip()]
-                    config_results: List[ConfigResult] = []
-
-                    iterator = (
-                        tqdm(lines, desc="Testing", leave=False, unit="cfg")
-                        if CONFIG.enable_url_testing
-                        else lines
-                    )
-                    for line in iterator:
-                        if all(
-                            [
-                                line.startswith(CONFIG.valid_prefixes),
-                                len(line) > 20,
-                                len(line) < 2000,
-                                len(config_results) < CONFIG.max_configs_per_source,
-                            ]
-                        ):
-                            config_hash = self.processor.create_semantic_hash(line)
-                            async with self.hash_lock:
-                                if config_hash in self.seen_hashes:
-                                    continue
-                                self.seen_hashes.add(config_hash)
-
-                            line = self.processor.apply_tuning(line)
-
-                            # Create config result
-                            host, port = self.processor.extract_host_port(line)
-                            protocol = self.processor.categorize_protocol(line)
-
-                            country = await self.processor.lookup_country(host) if host else None
-                            result = ConfigResult(
-                                config=line,
-                                protocol=protocol,
-                                host=host,
-                                port=port,
-                                source_url=url,
-                                country=country
-                            )
-                            
-                            # Test connection if enabled
-                            if CONFIG.enable_url_testing and host and port:
-                                ping_time = await self.processor.test_connection(host, port)
-                                result.ping_time = ping_time
-                                result.is_reachable = ping_time is not None
-                                if self.history_callback:
-                                    await self.history_callback(
-                                        config_hash,
-                                        result.is_reachable,
-                                        ping_time,
-                                    )
-                            
-                            config_results.append(result)
-
-                    if iterator is not lines and hasattr(iterator, "close"):
-                        iterator.close()
-
-                    return url, config_results
-                    
-            except (aiohttp.ClientError, asyncio.TimeoutError) as exc:
-                logging.debug("fetch_source error on %s: %s", url, exc)
-                if attempt < CONFIG.max_retries - 1:
-                    # Use a capped and jittered delay to reduce tail latency
-                    await asyncio.sleep(min(3, 1.5 + random.random()))
-        if use_temp:
-            await session.close()
-        return url, []
+# Reuse global regex list from output_writer
+EXCLUDE_REGEXES: List[re.Pattern] = OUTPUT_EXCLUDE_REGEXES
 
 # ============================================================================
 # MAIN PROCESSOR WITH UNIFIED FUNCTIONALITY
@@ -1204,158 +742,7 @@ class UltimateVPNMerger:
         start_time: float,
         prefix: str = "",
     ) -> None:
-        """Generate comprehensive output files with all formats."""
-        async with self._file_lock, self.seen_hashes_lock:
-            # Create output directory
-            output_dir = Path(CONFIG.output_dir)
-            output_dir.mkdir(exist_ok=True)
-        
-        # Extract configs for traditional outputs
-            configs = [result.config for result in results]
-        
-        # Raw text output
-            raw_file = output_dir / f"{prefix}vpn_subscription_raw.txt"
-            tmp_raw = raw_file.with_suffix('.tmp')
-            tmp_raw.write_text("\n".join(configs), encoding="utf-8")
-            tmp_raw.replace(raw_file)
-        
-            base64_file = output_dir / f"{prefix}vpn_subscription_base64.txt"
-            if CONFIG.write_base64:
-                base64_content = base64.b64encode("\n".join(configs).encode("utf-8")).decode("utf-8")
-                tmp_base64 = base64_file.with_suffix('.tmp')
-                tmp_base64.write_text(base64_content, encoding="utf-8")
-                tmp_base64.replace(base64_file)
-
-        # Enhanced CSV with comprehensive performance data
-            csv_file = output_dir / f"{prefix}vpn_detailed.csv"
-            if CONFIG.write_csv:
-                tmp_csv = csv_file.with_suffix('.tmp')
-                with open(tmp_csv, 'w', newline='', encoding='utf-8') as f:
-                    writer = csv.writer(f)
-                    writer.writerow([
-                        'config', 'protocol', 'host', 'port',
-                        'ping_ms', 'reachable', 'source_url', 'country'
-                    ])
-                    for result in results:
-                        ping_ms = round(result.ping_time * 1000, 2) if result.ping_time else None
-                        writer.writerow([
-                            result.config,
-                            result.protocol,
-                            result.host,
-                            result.port,
-                            ping_ms,
-                            result.is_reachable,
-                            result.source_url,
-                            result.country,
-                        ])
-                tmp_csv.replace(csv_file)
-        
-        # Comprehensive JSON report
-            report_file = output_dir / f"{prefix}vpn_report.json"
-            report = {
-            "generation_info": {
-                "timestamp_utc": datetime.now(timezone.utc).isoformat(),
-                "processing_time_seconds": round(time.time() - start_time, 2),
-                "script_version": "Unified & Polished Edition",
-                "url_testing_enabled": CONFIG.enable_url_testing,
-                "sorting_enabled": CONFIG.enable_sorting,
-            },
-            "statistics": stats,
-            "source_categories": {
-                "total_unique_sources": len(self.sources)
-            },
-            "output_files": {
-                "raw": str(raw_file),
-                **({"base64": str(base64_file)} if CONFIG.write_base64 else {}),
-                **({"detailed_csv": str(csv_file)} if CONFIG.write_csv else {}),
-                "json_report": str(report_file),
-                "singbox": str(output_dir / f"{prefix}vpn_singbox.json"),
-                "clash": str(output_dir / f"{prefix}clash.yaml"),
-                **(
-                    {"clash_proxies": str(output_dir / f"{prefix}vpn_clash_proxies.yaml")}
-                    if CONFIG.write_clash_proxies
-                    else {}
-                ),
-            },
-            "usage_instructions": {
-                "base64_subscription": "Copy content of base64 file as subscription URL",
-                "raw_subscription": "Host raw file and use URL as subscription link",
-                "csv_analysis": "Use CSV file for detailed analysis and custom filtering",
-                "clash_yaml": "Load clash.yaml in Clash Meta or Stash",
-                "clash_proxies_yaml": "Import vpn_clash_proxies.yaml as a simple provider",
-                "supported_clients": [
-                    "V2rayNG", "V2rayN", "Hiddify Next", "Shadowrocket",
-                    "NekoBox", "Clash Meta", "Sing-Box", "Streisand", "Karing"
-                ]
-            }
-        }
-        
-            tmp_report = report_file.with_suffix('.tmp')
-            tmp_report.write_text(json.dumps(report, indent=2, ensure_ascii=False), encoding="utf-8")
-            tmp_report.replace(report_file)
-
-        # Simple outbounds JSON
-            outbounds = []
-            for idx, r in enumerate(results):
-                tag = re.sub(r"[^A-Za-z0-9_-]+", "-", f"{r.protocol}-{idx}")
-                ob = {
-                    "type": r.protocol.lower(),
-                    "tag": tag,
-                    "server": r.host or "",
-                    "server_port": r.port,
-                    "raw": r.config
-                }
-                ob.update(self._parse_extra_params(r.config))
-                if r.country:
-                    ob["country"] = r.country
-                outbounds.append(ob)
-
-            singbox_file = output_dir / f"{prefix}vpn_singbox.json"
-            tmp_singbox = singbox_file.with_suffix('.tmp')
-            tmp_singbox.write_text(json.dumps({"outbounds": outbounds}, indent=2, ensure_ascii=False), encoding="utf-8")
-            tmp_singbox.replace(singbox_file)
-
-            if CONFIG.write_clash:
-                clash_yaml = self._results_to_clash_yaml(results)
-                clash_file = output_dir / f"{prefix}clash.yaml"
-                tmp_clash = clash_file.with_suffix('.tmp')
-                tmp_clash.write_text(clash_yaml, encoding="utf-8")
-                tmp_clash.replace(clash_file)
-
-            need_proxies = CONFIG.write_clash_proxies or CONFIG.surge_file or CONFIG.qx_file
-            proxies: List[Dict[str, Any]] = []
-            if need_proxies:
-                for idx, r in enumerate(results):
-                    proxy = config_to_clash_proxy(r.config, idx, r.protocol)
-                    if proxy:
-                        proxies.append(proxy)
-
-            if CONFIG.write_clash_proxies:
-                proxy_yaml = yaml.safe_dump({"proxies": proxies}, allow_unicode=True, sort_keys=False) if proxies else ""
-                proxies_file = output_dir / f"{prefix}vpn_clash_proxies.yaml"
-                tmp_proxies = proxies_file.with_suffix('.tmp')
-                tmp_proxies.write_text(proxy_yaml, encoding="utf-8")
-                tmp_proxies.replace(proxies_file)
-
-            if CONFIG.surge_file:
-                from .advanced_converters import generate_surge_conf
-                surge_content = generate_surge_conf(proxies)
-                surge_file = Path(CONFIG.surge_file)
-                if not surge_file.is_absolute():
-                    surge_file = output_dir / surge_file
-                tmp_surge = surge_file.with_suffix('.tmp')
-                tmp_surge.write_text(surge_content, encoding="utf-8")
-                tmp_surge.replace(surge_file)
-
-            if CONFIG.qx_file:
-                from .advanced_converters import generate_qx_conf
-                qx_content = generate_qx_conf(proxies)
-                qx_file = Path(CONFIG.qx_file)
-                if not qx_file.is_absolute():
-                    qx_file = output_dir / qx_file
-                tmp_qx = qx_file.with_suffix('.tmp')
-                tmp_qx.write_text(qx_content, encoding="utf-8")
-                tmp_qx.replace(qx_file)
+        await generate_comprehensive_outputs(self, results, stats, start_time, prefix)
 
 
     def _results_to_clash_yaml(self, results: List[ConfigResult]) -> str:
